Added Node.ReplaceChildren and Node.ReplaceChildrenWith
ReplaceChildren allows replacing of all nodes in a Node
with a new set of nodes.

This new set may contain already existing child nodes
and in that case the existing nodes will be moved into their new
position.

This method is useful when wanting to replace child nodes that have a
few changes.

For major changes a better approach would be to clear the
child nodes and add them again.